HiveMind/src/extractor/requirements.txt
google-cloud-storage
google-cloud-bigquery
google-api-python-client
google-auth
google-cloud-aiplatform
